+++ resources/lang/en/spec.php
@@ -100,6 +100,12 @@ return [
         'code' => '',
     ],
 
+    'resource_type_not_supported_by_relationship' => [
+        'title' => 'Not Supported',
+        'detail' => 'Resource type :type is not supported.',
+        'code' => '',
+    ],
+
     'resource_type_not_recognised' => [
         'title' => 'Not Supported',
         'detail' => 'Resource type :type is not recognised.',