+++ src/Translator.php
@@ -237,6 +237,23 @@ class Translator
     }
 
     /**
+     * Create an error for when the resource type is not supported by a relationship.
+     *
+     * @param string $type
+     * @param string $path
+     * @return Error
+     */
+    public function resourceTypeNotSupportedByRelationship(string $type, string $path = '/data'): Error
+    {
+        return Error::make()
+            ->setStatus(400)
+            ->setCode($this->trans('resource_type_not_supported_by_relationship', 'code'))
+            ->setTitle($this->trans('resource_type_not_supported_by_relationship', 'title'))
+            ->setDetail($this->trans('resource_type_not_supported_by_relationship', 'detail', compact('type')))
+            ->setSourcePointer($this->pointer($path, 'type'));
+    }
+
+    /**
      * Create an error for when a resource type is not recognised.
      *
      * @param string $type