+++ src/Values/ToMany.php
@@ -19,6 +19,8 @@ declare(strict_types=1);
 
 namespace LaravelJsonApi\Spec\Values;
 
+use LaravelJsonApi\Contracts\Schema\PolymorphicRelation;
+use LaravelJsonApi\Contracts\Schema\Relation;
 use LaravelJsonApi\Core\Document\ErrorList;
 use LaravelJsonApi\Spec\Factory;
 use LaravelJsonApi\Spec\Translator;
@@ -38,9 +40,9 @@ class ToMany extends Value
     private Factory $factory;
 
     /**
-     * @var string
+     * @var Relation
      */
-    private string $name;
+    private Relation $relation;
 
     /**
      * @var mixed
@@ -57,15 +59,15 @@ class ToMany extends Value
      *
      * @param Translator $translator
      * @param Factory $factory
-     * @param string $name
+     * @param Relation $relation
      * @param string $path
      * @param mixed $value
      */
-    public function __construct(Translator $translator, Factory $factory, string $name, string $path, $value)
+    public function __construct(Translator $translator, Factory $factory, Relation $relation, string $path, $value)
     {
         $this->translator = $translator;
         $this->factory = $factory;
-        $this->name = $name;
+        $this->relation = $relation;
         $this->path = rtrim($path, '/');
         $this->value = $value;
     }
@@ -80,9 +82,13 @@ class ToMany extends Value
         }
 
         if ($this->valid()) {
-            return $this->data = collect($this->value->data)
-                ->map(fn($value, $idx) => $this->factory->createIdentifierValue("{$this->path}/data/{$idx}", $value))
-                ->all();
+            return $this->data = collect($this->value->data)->map(function ($value, $idx) {
+                return $this->factory->createIdentifierValue(
+                    "{$this->path}/data/{$idx}",
+                    $value,
+                    $this->expected(),
+                );
+            })->all();
         }
 
         throw new LogicException('Invalid to-many relationship object.');
@@ -128,7 +134,7 @@ class ToMany extends Value
             return $errors->push($this->translator->fieldExpectsToMany(
                 $this->parent(),
                 $this->member() ?: 'data',
-                $this->name
+                $this->relation->name(),
             ));
         }
 
@@ -140,6 +146,20 @@ class ToMany extends Value
         }
 
         return $errors;
+    }
+
+    /**
+     * Get the expected resource types.
+     *
+     * @return array
+     */
+    private function expected(): array
+    {
+        if ($this->relation instanceof PolymorphicRelation) {
+            return $this->relation->inverseTypes();
+        }
+
+        return [$this->relation->inverse()];
     }
 
 }