+++ tests/Integration/TestCase.php
@@ -61,12 +61,14 @@ class TestCase extends BaseTestCase
 
     /**
      * @param string $name
+     * @param string $inverse
      * @return Relation
      */
-    protected function createToOne(string $name): Relation
+    protected function createToOne(string $name, string $inverse): Relation
     {
         $relation = $this->createMock(Relation::class);
         $relation->method('name')->willReturn($name);
+        $relation->method('inverse')->willReturn($inverse);
         $relation->method('toOne')->willReturn(true);
         $relation->method('toMany')->willReturn(false);
 
@@ -75,12 +77,14 @@ class TestCase extends BaseTestCase
 
     /**
      * @param string $name
+     * @param string $inverse
      * @return Relation
      */
-    protected function createToMany(string $name): Relation
+    protected function createToMany(string $name, string $inverse): Relation
     {
         $relation = $this->createMock(Relation::class);
         $relation->method('name')->willReturn($name);
+        $relation->method('inverse')->willReturn($inverse);
         $relation->method('toOne')->willReturn(false);
         $relation->method('toMany')->willReturn(true);