+++ tests/Integration/ToManyTest.php
@@ -37,7 +37,7 @@ class ToManyTest extends TestCase
         $spec->method('exists')->willReturnCallback(fn($type, $id) => '999' !== $id);
         $spec->method('fields')->willReturnMap([
             ['posts', [
-                $this->createToMany('tags'),
+                $this->createToMany('tags', 'tags'),
             ]],
             ['tags', [
                 $this->createAttribute('name'),
@@ -121,6 +121,19 @@ class ToManyTest extends TestCase
                 [
                     'title' => 'Non-Compliant JSON:API Document',
                     'detail' => "The member type cannot be empty.",
+                    'status' => '400',
+                    'source' => ['pointer' => '/data/0/type'],
+                ],
+            ],
+            'data.type:not supported' => [
+                [
+                    'data' => [
+                        ['type' => 'posts', 'id' => '1'],
+                    ],
+                ],
+                [
+                    'title' => 'Not Supported',
+                    'detail' => "Resource type posts is not supported.",
                     'status' => '400',
                     'source' => ['pointer' => '/data/0/type'],
                 ],